requirements.txt
matplotlib==3.8.2
numpy==2.0.1
opencv_python==4.9.0.80
Pillow==9.4.0
Pillow==10.4.0
scipy==1.14.0
termcolor==2.4.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.65.0
